feat: 🍻 [WIP] physics physics physics

+++ internal/domain/entity/vector.go
@@ -1,7 +1,7 @@
 package entity
 
 type Vector3 struct {
-	X float32
-	Y float32
-	Z float32
+	X float64
+	Y float64
+	Z float64
 }
